Refactor (move homepage into an own file)
