+++ lib/homepage.dart
@@ -1,0 +1,41 @@
+import 'package:flutter/material.dart';
+
+import 'customBottomAppBar.dart';
+
+class HomePage extends StatefulWidget {
+  const HomePage({Key? key, required this.title}) : super(key: key);
+
+  final String title;
+
+  @override
+  State<HomePage> createState() => _HomePageState();
+}
+
+class _HomePageState extends State<HomePage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(widget.title),
+        actions: [
+          IconButton(
+            tooltip: 'More options',
+            icon: const Icon(Icons.more_vert),
+            onPressed: () {
+              print('More options clicked');
+            }, // TODO: Add menu (Settings, About, etc)
+          ),
+        ],
+      ),
+      floatingActionButton: FloatingActionButton(
+        child: const Icon(Icons.play_arrow),
+        onPressed: () {
+          print('Start clicked');
+        },
+        tooltip: 'Start',
+      ),
+      floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
+      bottomNavigationBar: const CustomBottomAppBar(),
+    );
+  }
+}
